requirements.txt
pandas
python-whois
tldextract
aiohttp
playwright
dnspython